FEAT: 상세페이지 구현
